bugfix: Restriction must inherit Op.dtype

+++ pylops/signalprocessing/Sliding2D.py
@@ -135,7 +135,9 @@ def Sliding2D(Op, dims, dimsd, nwin, nover, tapertype="hanning", design=False):
 
     combining = HStack(
         [
-            Restriction(np.prod(dimsd), range(win_in, win_end), dims=dimsd).H
+            Restriction(
+                np.prod(dimsd), range(win_in, win_end), dims=dimsd, dtype=Op.dtype
+            ).H
             for win_in, win_end in zip(dwin_ins, dwin_ends)
         ]
     )
